StateMachine fixes retained callback lifetime and adds async lifecycle demo
Completes the SM-002-R1A callback-lifetime investigation and follow-up hardening.

The debug-heap failure was traced to an invalid destruction test: startup settlement
destroyed the StateMachine, then the test used its saved raw pointer. The retained
callback observer itself was proven independently and retained using U++-native
`One<Pte<T>>` / `Ptr<T>` semantics.

## Core changes

- Uses a machine-owned active `One<Operation>`.
- `Operation` derives from `Pte<Operation>`.
- Retained `done(bool)` callbacks carry only `Ptr<Operation>` and callback phase.
- Stale callbacks after cancellation, settlement, reset, clear, or destruction are inert.
- Post-hook lifetime guards prevent member access after re-entrant destruction.
- Corrects destruction tests and adds focused coverage for destruction from:
  - initial OnEnter;
  - source OnExit;
  - target OnEnter;
  - WhenTransitionFinished;
  - OnAfter;
  - WhenTransitionSettled.

## Verification

- Added isolated U++ `One/Pte/Ptr/Function` ownership proof, including the by-value
  callback-passing shape used by StateMachine.
- Debug-heap core suite: **218/218 passed**.
- StateMachineGuiTest: builds.
- StateMachineVisualizer: builds and launches.

## Visualizer / UI follow-up

- Adds automatic lifecycle demonstration, toggle, interval slider, and Async Monitor
  route visualization.
- Async Monitor samples Quality Check units and visibly routes them to Packaging or
  Quality Review.
- Fixes `UiTitleCard::CancelMode` nested capture release recursion found while
  clicking Visualizer nodes.

+++ examples/StateMachineGuiTest/main.cpp
@@ -233,12 +233,12 @@ struct TestRunner : TopWindow {
         sm.AddTransition({"go", "A", "B"});
         sm.WhenTransitionSettled = [&](const TransitionResult& result) {
             ++settlements;
-            logDisplay.Log(Format("  settled: sequence=%llu operation=%d outcome=%d", (unsigned long long)result.sequence, (int)result.operation, (int)result.outcome));
+            logDisplay.Log(Format("  settled: sequence=%lld operation=%d outcome=%d", (int64)result.sequence, (int)result.operation, (int)result.outcome));
         };
         sm.Start();
         const int history_before = sm.GetHistoryCount();
         sm.TriggerEvent("go");
-        logDisplay.Log(Format("  active sequence=%llu operation=%d", (unsigned long long)sm.GetActiveTransitionSequence(), (int)TransitionOperationKind::Transition));
+        logDisplay.Log(Format("  active sequence=%lld operation=%d", (int64)sm.GetActiveTransitionSequence(), (int)TransitionOperationKind::Transition));
         ASSERT(sm.CancelActiveTransition());
         ASSERT(sm.GetLastTransitionOutcome() == TransitionOutcome::Cancelled);
         ASSERT(sm.GetCurrent() == "A" && sm.GetHistoryCount() == history_before);
